ml/cloud_model/training/videomae_train/requirements.txt

opencv-python
numpy
pyyaml
tqdm
scikit-learn
transformers
accelerate

